backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
scikit-learn  
joblib       
pandas      